Fix original content display
Before, when an article was containing a picture gallery and the configuration was
hiding the original content, the gallery was not displayed because it was considered
as original content even though it was added by the insert transformer.
Now, only the original content of the article is stripped before being displayed.