Fix display image option
Before, galleries were shown at all time even when the display image option was set
to false. This was a side effect of the preprocessing that does not (and will not)
take that setting into account.
Now, the preprocessed content is displayed only when the display image option is set
to true. The validation is simple because only images are added during preprocessing.