Add url validation before generating video tags
Before, the process was adding both a source for webm and mp4 videos. But this
was an issue since not every source provide the 2 formats. Thus making the
rendering quite unusable when in third-party tools.
Now, the mp4 format is preferred. On top of that the URL is checked before being
added to the video tag. It's a bit longer to process but it's much safer since
all URL are validated beforehand.

See #5